dependent_code/requirements.txt
requests
beautifulsoup4
pandas
fastapi
pydantic
uvicorn
streamlit
matplotlib
seaborn
tqdm
pytest
boto3
python-dotenv
psycopg2-binary
redis
python-dateutil
great_expectations==0.18.19
transformers
torch
scikit-learn
yfinance
pymongo
python-jose[cryptography]
lxml
numpy<2
prometheus_client
google-genai
celery[redis]
flower
passlib[bcrypt]
bcrypt==4.0.1
